error handling

+++ bedrock/src/nitro_enclave/types.rs
@@ -24,7 +24,7 @@ pub enum EnclaveAttestationError {
     #[error("PCR{pcr_index} value not trusted: {actual}")]
     CodeUntrusted {
         /// The index of the PCR value that failed validation
-        pcr_index: usize,
+        pcr_index: u32,
         /// The actual value of the PCR that failed validation
         actual: String,
     },
@@ -59,7 +59,7 @@ pub type EnclaveAttestationResult<T, E = EnclaveAttestationError> = Result<T, E>
 pub struct PcrConfiguration {
     /// The index of the PCR{index} value
     /// eg. 0, 1, 2, 3, 4, 8
-    pub index: usize,
+    pub index: u32,
     /// The expected value of the PCR
     pub expected_value: Vec<u8>,
     /// The description of the PCR
